BSgenome.Dpulex.JGI.dpulex {BSgenome.Dpulex.JGI.dpulex}	R Documentation

Full genome sequences for Daphnia pulex TCO (JGI v1.0)

Description

Full genome sequence for Daphnia pulex (clone TCO) as provided by JGI and Indiana University

Note

This BSgenome data package was made from the following source data files:

-- information not available --

  
See ?BSgenomeForge and the BSgenomeForge vignette (vignette("BSgenomeForge")) in the BSgenome software package for how to make a BSgenome data package.
Author(s)

The Bioconductor Dev Team

See Also

BSgenome objects and the available.genomes function in the BSgenome software package.

DNAString objects in the Biostrings package.

The BSgenomeForge vignette (vignette("BSgenomeForge")) in the BSgenome software package for how to make a BSgenome data package.

Examples

BSgenome.Dpulex.JGI.dpulex
genome <- BSgenome.Dpulex.JGI.dpulex
seqlengths(genome)

## See the GenomeSearching vignette in the BSgenome software
## package for some examples of genome-wide motif searching using
## Biostrings and the BSgenome data packages:
if (interactive())
    vignette("GenomeSearching", package="BSgenome")

[Package BSgenome.Dpulex.JGI.dpulex version 1.1 Index]